Arreglo errores archivos y separacion
