+++ PruebaDeNivelNasa/Models/DTOS/AsteroidDTO.cs
@@ -1,24 +1,5 @@
-﻿namespace PruebaDeNivelNasa.Models
+﻿namespace PruebaDeNivelNasa.Models.DTOS
 {
-    //TODO: debe haber únicament euna clase por archivo
-    //TODO: deben estar separados los modelos de obtención/parseo de la API de los DTOS, por ejemplo con carpetas
-    public class ResponseDTO
-    {
-        public List<AsteroidDTO> List { get; set; } = new List<AsteroidDTO>();
-        public override bool Equals(object obj)
-        {
-            if (obj == null) return false;
-            if (ReferenceEquals(this, obj)) return true;
-            if (obj.GetType() != GetType()) return false;
-            ResponseDTO other = (ResponseDTO)obj;
-            return List.SequenceEqual(other.List);
-        }
-        public override int GetHashCode()
-        {
-            return base.GetHashCode();
-        }
-    }
-
     public class AsteroidDTO
     {
         public string Nombre { get; set; }
