+++ PruebaDeNivelNasa/Models/ResultAPI/Close_approach_data.cs
@@ -1,0 +1,9 @@
+﻿namespace PruebaDeNivelNasa.Models.ResultAPI
+{
+    public class Close_approach_data
+    {
+        public DateOnly close_approach_date { get; set; } = new DateOnly();
+        public Relative_velocity relative_velocity { get; set; } = new Relative_velocity();
+        public string orbiting_body { get; set; }
+    }
+}
