+++ PruebaDeNivelNasa/Models/ResultAPI/Estimated.cs
@@ -1,0 +1,8 @@
+﻿namespace PruebaDeNivelNasa.Models.ResultAPI
+{
+    public class Estimated
+    {
+        public decimal estimated_diameter_min { get; set; }
+        public decimal estimated_diameter_max { get; set; }
+    }
+}
