+++ PruebaDeNivelNasa/Models/ResultAPI/Estimated_diameter.cs
@@ -1,0 +1,7 @@
+﻿namespace PruebaDeNivelNasa.Models.ResultAPI
+{
+    public class Estimated_diameter
+    {
+        public Estimated kilometers { get; set; }
+    }
+}
