+++ PruebaDeNivelNasa/Models/ResultAPI/Links.cs
@@ -1,0 +1,7 @@
+﻿namespace PruebaDeNivelNasa.Models.ResultAPI
+{
+    public class Links
+    {
+        public string self { get; set; }
+    }
+}
