+++ PruebaDeNivelNasa/Models/ResultAPI/Relative_velocity.cs
@@ -1,0 +1,7 @@
+﻿namespace PruebaDeNivelNasa.Models.ResultAPI
+{
+    public class Relative_velocity
+    {
+        public decimal kilometers_per_hour { get; set; }
+    }
+}
