+++ PruebaDeNivelNasa/Services/Interfaces/INasaService.cs
@@ -1,6 +1,7 @@
-﻿using PruebaDeNivelNasa.Models;
+﻿using PruebaDeNivelNasa.Models.DTOS;
+using PruebaDeNivelNasa.Models.ResultAPI;
 
-namespace PruebaDeNivelNasa.Services
+namespace PruebaDeNivelNasa.Services.Interfaces
 {
     public interface INasaService
     {
