+++ Test/Test.csproj
@@ -21,4 +21,8 @@
     <ProjectReference Include="..\PruebaDeNivelNasa\PruebaDeNivelNasa.csproj" />
   </ItemGroup>
 
+  <ItemGroup>
+    <Folder Include="Mock\" />
+  </ItemGroup>
+
 </Project>
